Remove ScalaJS support
