--- quill-cassandra-lagom/src/test/resources/application.conf
@@ -1,18 +1,0 @@
-testStreamDB.keyspace=quill_test
-testStreamDB.preparedStatementCacheSize=1000
-testStreamDB.session.contactPoint=127.0.0.1
-testStreamDB.session.contactPoint=${?CASSANDRA_HOST}
-testStreamDB.session.port=9042
-testStreamDB.session.port=${?CASSANDRA_PORT}
-testStreamDB.session.queryOptions.fetchSize=1
-testStreamDB.session.queryOptions.consistencyLevel=LOCAL_QUORUM
-
-cassandra-journal.keyspace = ${testStreamDB.keyspace}
-cassandra-snapshot-store.keyspace = ${testStreamDB.keyspace}
-lagom.persistence.read-side.cassandra.keyspace = ${testStreamDB.keyspace}
-
-lagom.services {
-  cas_native = "tcp://"${testStreamDB.session.contactPoint}":"${testStreamDB.session.port}
-}
-
-akka.discovery.method = "akka-dns"
