--- quill-cassandra-lagom/src/test/scala/io/getquill/context/cassandra/lagom/package.scala
@@ -1,8 +1,0 @@
-package io.getquill.context.cassandra
-
-import io.getquill.{CassandraLagomAsyncContext, Literal}
-import utils._
-
-package object lagom {
-  lazy val testLagomAsyncDB = new CassandraLagomAsyncContext(Literal, cassandraSession) with CassandraTestEntities
-}
