--- quill-cassandra-lagom/src/test/scala/io/getquill/context/cassandra/streaming/package.scala
@@ -1,8 +1,0 @@
-package io.getquill.context.cassandra
-
-import io.getquill.{CassandraLagomStreamContext, Literal}
-import utils._
-
-package object streaming {
-  lazy val testStreamDB = new CassandraLagomStreamContext(Literal, cassandraSession) with CassandraTestEntities
-}
