--- quill-cassandra-lagom/src/test/scala/io/getquill/context/cassandra/utils/DummyService.scala
@@ -1,10 +1,0 @@
-package io.getquill.context.cassandra.utils
-
-import com.lightbend.lagom.scaladsl.api.{Descriptor, Service}
-
-class DummyService extends Service {
-  override def descriptor: Descriptor = {
-    import Service._
-    named("dummy")
-  }
-}
