--- quill-core/js/src/main/scala/io/getquill/log/ContextLog.scala
@@ -1,5 +1,0 @@
-package io.getquill.log
-
-object ContextLog {
-  def apply(str: String) = println(str)
-}
