+++ quill-core/src/test/scala/io/getquill/util/LogToFileSpec.scala
@@ -1,7 +1,7 @@
 package io.getquill.util
 
 import io.getquill.base.Spec
-import io.getquill.util.Messages.LogToFile
+
 import scala.io.Source
 
 class LogToFileSpec extends Spec {
